Final feature 7 code , fully working.
I realised I had made the code too complex by calling the employeesListToJson method into the EmployeeDaoInterface and MySqlEmployeeDao classes which was overriding my method in the jsonConverter class. I instead created a jsonConverter object in the MainApp and called the method in this way, and now it is working as it should.