Final changes
Feature 13 added - from the shared github. Was having an issue with the employees each printing out twice in the display after delete was called, so I copied over the code that was in the shared and it magically worked so that issue is fixed. TODO's are done - this was some validation.